pyproject.toml
[project]
name = "gemini-agent"
version = "0.1.0"
description = "A portable Gemini LangChain implementation"
authors = [
    {name = "Brady Anderson", email = "brady@bradyanderson.tech"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "langchain>=0.3.18",
    "google-generativeai>=0.8.4",
    "python-dotenv>=1.0.0",
    "langchain-google-genai>=2.0.9",
    "langchain-core>=0.3.35",
    "langgraph (>=0.2.72,<0.3.0)"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
gemini-chat = "gemini_agent.main:main"